- OkiAdpcm: add gain parameter to DialogicAdpcmSamp and amplify samples on conversion. - CPS1Instr: improve handling of CPS1 4.25 driver instrument params - CPSSeq: add instrument transposition table for CPS1 YM2151 tracks. - CPSSeq: fix note on/off order for CPS1 (will revisit portamento to follow the AkaoaSeq approach down the road) - update mame_roms.xml - update ymmy

+++ src/main/formats/CPS/CPSTrackV1.h
@@ -19,15 +19,15 @@ private:
   void calculateAndAddPortamentoTimeNoItem(int8_t noteDistance);
 
   CPSSynth channelSynth;
-  uint32_t noteDuration;
-  int8_t key;
+  u8 noteDuration;
+  s8 key;
   bool bPrevNoteTie;
-  uint8_t prevTieNote;
-  uint8_t curDeltaTable;
-  uint8_t noteState;
-  uint8_t bank;
-  uint8_t loop[4];
-  uint32_t loopOffset[4];    //used for detecting infinite loops
-  int16_t portamentoCentsPerSec;
-  uint16_t prevPortamentoDuration;
+  u8 prevTieNote;
+  u8 curDeltaTable;
+  u8 noteState;
+  u8 bank;
+  u8 loop[4];
+  u32 loopOffset[4];    //used for detecting infinite loops
+  s16 portamentoCentsPerSec;
+  u16 prevPortamentoDuration;
 };
